reorg keystore
