+++ src/Keystore/constants.ts
@@ -1,0 +1,6 @@
+export const KEYSTORE_VERSION = '6.0';
+
+export const ITERATIONS_V2 = 100000;
+export const ITERATIONS_V3 = 200000; // and any version above
+
+export const SUPPORTED_VERSION = ['2.0', '3.0', '4.0', '5.0', '6.0'];
