+++ src/Keystore/index.ts
@@ -1,2 +1,7 @@
-export * from './keystore';
+export * from './constants';
+export * from './CryptoHelpers';
+export * from './extractKeys';
+export * from './makeFile';
+export * from './readFile';
+export * from './utils';
 export * from './types';
