+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './fee_utils';
 export * from './price_utils';
 export * from './waitTxUtils';
 export * from './sleep';
+export * from './bintools';
